+++ src/components/Heading/Heading.tsx
@@ -1,27 +1,37 @@
 import * as React from 'react'
-import { css, ResponsiveFontSize, shouldTruncate, styled } from '../../style'
+import {
+  getTextTransform,
+  ResponsiveFontSize,
+  shouldTruncate,
+  styled,
+  TextTransforms,
+  textVariant,
+  TextVariants,
+} from '../../style'
 import { Box, BoxPropsWithout } from '../Box'
 
 export type HeadingLevels = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'
 export type HeadingTextTransforms = 'caps' | 'lower' | 'none' | 'upper'
 
 export interface HeadingProps
-  extends BoxPropsWithout<HTMLHeadingElement, 'color' | 'truncate'> {
+  extends BoxPropsWithout<HTMLHeadingElement, 'truncate'> {
   /** Heading level from h1-h6 */
   is?: HeadingLevels
   /** Text tranform  */
-  transform?: HeadingTextTransforms
+  textTransform?: TextTransforms
   /** Truncate heading text */
   truncate?: boolean
   /** Custom css class */
   className?: string
+  /** Adjust style of text with more meaning by using an intent */
+  variant?: TextVariants
 }
 
 const InternalHeading: React.FC<HeadingProps> = ({
   fontSize,
   fontWeight,
   lineHeight,
-  transform,
+  textTransform,
   truncate,
   is,
   ...props
@@ -42,31 +52,10 @@ const HeadingFactory = React.forwardRef((props: HeadingProps, ref) => (
 ))
 
 export const Heading = styled(HeadingFactory)`
-  ${props => textTransform(props.transform)};
+  ${props => getTextTransform(props.textTransform)};
   ${props => shouldTruncate(props.truncate || false)};
+  ${props => textVariant(props.theme, props.variant)};
 `
-
-const textTransform = (transform: HeadingTextTransforms | undefined) => {
-  switch (transform) {
-    case 'upper':
-      return css`
-        text-transform: uppercase;
-      `
-    case 'lower':
-      return css`
-        text-transform: lowercase;
-      `
-    case 'caps':
-      return css`
-        text-transform: capitalize;
-      `
-    case 'none':
-    default:
-      return css`
-        text-transform: none;
-      `
-  }
-}
 
 const headingLevelSize = (is?: HeadingLevels) => {
   switch (is) {